+++ spec/defines/init_spec.rb
@@ -26,7 +26,7 @@ describe 'puppet_device' do
     it { is_expected.to contain_class('puppet_device::conf') }
     it { is_expected.to contain_class('puppet_device::fact') }
     it { is_expected.to contain_class('puppet_device::run') }
-    it { is_expected.to contain_exec('run puppet_device').with_command("/opt/puppetlabs/puppet/bin/puppet device --user=root --waitforcert 0") }
+    it { is_expected.to contain_exec('run puppet_device').with_command("/opt/puppetlabs/puppet/bin/puppet device --user=root --waitforcert=0") }
   end
 
   context 'on Windows, with values for all parameters' do
@@ -53,7 +53,7 @@ describe 'puppet_device' do
     it { is_expected.to contain_class('puppet_device::conf') }
     it { is_expected.to contain_class('puppet_device::fact') }
     it { is_expected.to contain_class('puppet_device::run') }
-    it { is_expected.to contain_exec('run puppet_device').with_command("\"C:/Program Files/Puppet Labs/Puppet/bin/puppet\" device --user=root --waitforcert 0") }
+    it { is_expected.to contain_exec('run puppet_device').with_command("\"C:/Program Files/Puppet Labs/Puppet/bin/puppet\" device --user=root --waitforcert=0") }
   end
 
   context 'running Puppet 5.0, on Linux, with values for all parameters' do
@@ -81,7 +81,7 @@ describe 'puppet_device' do
     it { is_expected.to contain_class('puppet_device::conf') }
     it { is_expected.to contain_class('puppet_device::fact') }
 
-    it { is_expected.to contain_exec('init puppet_device target bigip').with_command("/opt/puppetlabs/puppet/bin/puppet device --target bigip --user=root --waitforcert 0") }
-    it { is_expected.to contain_exec('run puppet_device target bigip').with_command("/opt/puppetlabs/puppet/bin/puppet device --target bigip --waitforcert 0") }
+    it { is_expected.to contain_exec('init puppet_device target bigip').with_command("/opt/puppetlabs/puppet/bin/puppet device --target bigip --user=root --waitforcert=0") }
+    it { is_expected.to contain_exec('run puppet_device target bigip').with_command("/opt/puppetlabs/puppet/bin/puppet device --target bigip --user=root --waitforcert=0") }
   end
 end